api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
psycopg2-binary==2.9.9
minio==7.2.7
pyjwt==2.8.0
python-multipart==0.0.9
